Add support for inlining struct fields.
msgpack encoder differ in behaviour from the json one - the latter inlines
fields of an embedded struct as they were part of the parent struct.

While we can't change the behaviour due to compatibility reasons we can
introduce `inline` opt, which will make the encoding similar in
behaviour to the json one.