Fallback to json struct tag if no msgpack tag is present
This change is related to #149 and adds fallback option to `json:`
struct tags if no `msgpack:` tags are found. This is useful if you want
to serialize a generate struct that already has the `json:` tags and you
want to `msgpack` it as well.

To enable the `json:` tag fallback set `Encoder.UseJSONTag(true)`.