Add Canonical option to encoder
Encoded values can be used as keys by external systems.
It is the case when the encoded value is used for caching or as a
primary key in a DB.
In order to be able to use msgpack encoded value as a key, we must
ensure that the generated byte array is always the same for the same
input.
This is important for maps and structs.
For maps, as they are unordered, ordering must be added to the resulting
encoded value.
For slices, as they are serialized as maps, they must conform to the
same ordering as maps.

This change is adding a Canonical option to the Encoder.
When this option is enabled, the Encoder ensures ordering or maps and
structs so they can be used as keys.