Omit empty simple sturcts
Currently, we only omit default values for primitives, interfaces and,
pointers. This makes it difficult to determine if a primitive was
present in the unpacked message.

Whilst unpacking you can determine if a value is present by implementing
something similar to standard lib's database/sql.NullInt64 (or any of
the other Null* types), but then the "Valid" field will be included when
you marshal the struct.

To resovle this, we must check if the given struct is simple enough
(meaning it doesn't have any hidden fields) to compare to a
zero-initialized version of the struct. If it is simple enough, and the
provided struct is zero-initialized, and the struct is tagged with
"omitempty" we can exclude it from the resulting bytes.